requirements.txt
opencv-python==4.8.1.78
numpy==1.26.0
Pillow==10.0.1
scikit-learn==1.3.1
joblib==1.3.2
matplotlib==3.8.2